library-event-consumer/src/main/java/com/learnkafka/libraryeventconsumer/consumer/LibraryEventsConsumer.java
package com.learnkafka.libraryeventconsumer.consumer;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.learnkafka.libraryeventconsumer.service.LibraryEventsService;
import lombok.RequiredArgsConstructor;
import lombok.extern.log4j.Log4j2;
import org.apache.kafka.clients.consumer.ConsumerRecord;
import org.springframework.kafka.annotation.KafkaListener;
import org.springframework.stereotype.Component;

@Component
@Log4j2
@RequiredArgsConstructor
public class LibraryEventsConsumer {

    private final LibraryEventsService libraryEventsService;

    @KafkaListener(topics = "library-events")
    public void onMessage(ConsumerRecord<Integer, String> consumerRecord) throws JsonProcessingException {

        log.info("Message Received {}", consumerRecord);
        libraryEventsService.processLibraryEvent(consumerRecord);
    }

}
